server/requirements.txt
flask
pinecone-client
openai
flask-cors